Experimental Java to Bedrock Conversion
Haven't done any validity checks yet, so some textures may be inaccurate. I wanted to add them to the repo initially just to have something to start with.

Big thanks to Ewan Howell for the Resource Pack Converter script! It makes things so much easier, there would be a ton more files to check and rename, and that made it so much easier to at least get started.
https://github.com/ewanhowell5195/resourcePackConverter